+++ projects/demo/src/pages/components/input-number/examples/12/intl-pattern.ts
@@ -1,0 +1,26 @@
+import {inject, LOCALE_ID} from '@angular/core';
+
+export function intlThousandSeparatorPattern(
+    locale: string = inject(LOCALE_ID),
+): (digits: string) => readonly string[] {
+    const formatter = new Intl.NumberFormat(locale, {maximumFractionDigits: 0});
+
+    return (digits) => {
+        if (!digits) {
+            return [];
+        }
+
+        let position = 0;
+
+        return formatter
+            .formatToParts(BigInt(`1${'0'.repeat(digits.length - 1)}`))
+            .filter(({type}) => type === 'integer')
+            .map(({value}) => {
+                const group = digits.slice(position, position + value.length);
+
+                position += value.length;
+
+                return group;
+            });
+    };
+}
